Get rid of unused references
