+++ src/ScaffoldCollection.php
@@ -4,14 +4,6 @@ declare(strict_types = 1);
 
 namespace Grasmash\ComposerScaffold;
 
-use Composer\Package\PackageInterface;
-use Composer\Script\Event;
-use Composer\Composer;
-use Composer\EventDispatcher\EventDispatcher;
-use Composer\IO\IOInterface;
-use Composer\Util\Filesystem;
-use Symfony\Component\Filesystem\Filesystem as SymfonyFilesystem;
-
 /**
  * ScaffoldCollection keeps track of the collection of files to be scaffolded.
  */
