+++ src/ScaffoldFileInfo.php
@@ -4,13 +4,7 @@ declare(strict_types = 1);
 
 namespace Grasmash\ComposerScaffold;
 
-use Composer\Package\PackageInterface;
-use Composer\Script\Event;
-use Composer\Composer;
-use Composer\EventDispatcher\EventDispatcher;
 use Composer\IO\IOInterface;
-use Composer\Util\Filesystem;
-use Symfony\Component\Filesystem\Filesystem as SymfonyFilesystem;
 use Grasmash\ComposerScaffold\Operations\ScaffoldOperationInterface;
 
 /**
